c/openmp_matrix: fix signedness change in substraction
clang complained:

openmp_matrix.c:31:16: error: implicit conversion changes signedness:
'ssize_t' (aka 'long') to 'unsigned long' [-Werror,-Wsign-conversion]
        len -= count;
            ~~ ^~~~~

+++ c/openmp_matrix.c
@@ -28,7 +28,7 @@ static void print_nobuf(const char *string)
             break;
         }
         string += count;
-        len -= count;
+        len -= (size_t)count;
     }
 }
